+++ numsigned.go
@@ -82,19 +82,19 @@ func (b *SignedBuilder[T]) SeeAlso(i Input, options ...SeeAlsoOption) *SignedBui
 // Required completes the build process and registers a required variable with
 // Ferrite's validation system.
 func (b *SignedBuilder[T]) Required(options ...RequiredOption) Required[T] {
-	return required(b.schema, &b.spec, options)
+	return required(b.schema, &b.spec, options...)
 }
 
 // Optional completes the build process and registers an optional variable with
 // Ferrite's validation system.
 func (b *SignedBuilder[T]) Optional(options ...OptionalOption) Optional[T] {
-	return optional(b.schema, &b.spec, options)
+	return optional(b.schema, &b.spec, options...)
 }
 
 // Deprecated completes the build process and registers a deprecated variable
 // with Ferrite's validation system.
 func (b *SignedBuilder[T]) Deprecated(options ...DeprecatedOption) Deprecated[T] {
-	return deprecated(b.schema, &b.spec, options)
+	return deprecated(b.schema, &b.spec, options...)
 }
 
 type signedMarshaler[T constraints.Signed] struct{}